requirements.txt
requests
python-telegram-bot
pymongo
matplotlib
